+++ Source/LidarRays/Vehicle/Models/UnicycleModel.h
@@ -11,12 +11,12 @@ public:
 	void initModel() override;
 	void closeModel() override;
 
-	void requestControl(std::map<std::string, float> controlRequest) override;
-	void executeModel(float DeltaTime) override;
+	void requestControl(std::map<std::string, double> controlRequest) override;
+	void executeModel(double DeltaTime) override;
 
-	std::map<std::string, float> controlsToModel(const std::map<std::string, float>& inControl) const override;
-	std::map<std::string, float> statesToModel(const std::map<std::string, float>& inState) const override;
+	std::map<std::string, double> controlsToModel(const std::map<std::string, double>& inControl) const override;
+	std::map<std::string, double> statesToModel(const std::map<std::string, double>& inState) const override;
 
-	std::map<std::string, float> controlsToWorld(const std::map<std::string, float>& inControl) const override;
-	std::map<std::string, float> statesToWorld(const std::map<std::string, float>& inState) const override;
+	std::map<std::string, double> controlsToWorld(const std::map<std::string, double>& inControl) const override;
+	std::map<std::string, double> statesToWorld(const std::map<std::string, double>& inState) const override;
 };
